CATROID-1593 Fix error 'Could not launch activity' in Espresso tests
Android OS can only start activities if they are listed in the app's
Manifest. Activities created for testing purposes were therefore
moved into the debug source set and listed in the corresponding debug
Manifest. When the app us built for testing those activities will be
available but not when the app is built for release.
