+++ catroid/src/androidTest/java/org/catrobat/catroid/uiespresso/ui/activity/ConnectBluetoothDeviceActivityTest.kt
@@ -1,6 +1,6 @@
 /*
  * Catroid: An on-device visual programming system for Android devices
- * Copyright (C) 2010-2022 The Catrobat Team
+ * Copyright (C) 2010-2024 The Catrobat Team
  * (<http://developer.catrobat.org/credits>)
  *
  * This program is free software: you can redistribute it and/or modify
@@ -33,8 +33,8 @@ import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import org.catrobat.catroid.ui.ConnectBluetoothDeviceTestActivity
 import org.catrobat.catroid.R
-import org.catrobat.catroid.bluetooth.ConnectBluetoothDeviceActivity
 import org.catrobat.catroid.bluetooth.ConnectBluetoothDeviceActivity.DEVICE_TO_CONNECT
 import org.catrobat.catroid.bluetooth.base.BluetoothDevice.MULTIPLAYER
 import org.catrobat.catroid.common.SharedPreferenceKeys.SHOW_MULTIPLAYER_BLUETOOTH_DIALOG_KEY
@@ -118,10 +118,5 @@ class ConnectBluetoothDeviceActivityTest {
             .check(matches(withDrawable(R.drawable.ic_search)))
             .perform(click())
             .check(matches(withDrawable(R.drawable.ic_close)))
-    }
-
-    class ConnectBluetoothDeviceTestActivity : ConnectBluetoothDeviceActivity() {
-        override fun initBluetooth() = Unit
-        override fun doDiscovery() = Unit
     }
 }
