+++ catroid/src/debug/AndroidManifest.xml
@@ -91,6 +91,11 @@
         tools:replace="android:label,android:allowBackup,android:name"
         android:usesCleartextTraffic="true">
 
+    <activity
+        android:name=".ui.ConnectBluetoothDeviceTestActivity" />
+
+    <activity
+        android:name=".ui.ProjectUploadTestActivity" />
 <!--
         <activity
             android:name=".ui.SettingsActivity"
