+++ catroid/src/main/AndroidManifest.xml
@@ -79,9 +79,10 @@
 
     <uses-feature
         android:glEsVersion="0x00020000"
-        android:required="true" /><uses-feature
-    android:name="android.hardware.camera"
-    android:required="false" />
+        android:required="true" />
+    <uses-feature
+        android:name="android.hardware.camera"
+        android:required="false" />
 
     <application
         android:name=".CatroidApplication"
